requirements.txt
langchain-chroma
langchain
langchain_core
langchain_community
chromadb>=0.4.4
google-generativeai
streamlit
python-dotenv
tqdm
requests
pymupdf

